+++ server/resources/mixins/timestamp_mixin.py
@@ -1,0 +1,6 @@
+from datetime import datetime
+
+
+class TimestampMixin:
+    created_at: datetime
+    updated_at: datetime
