+++ server/resources/user_resource.py
@@ -2,7 +2,9 @@ from typing import Self
 
 from pydantic import model_validator
 
-from server.resources.base_resource import AllOptionalMetaclass, BaseResource
+from server.resources.base_resource import BaseResource
+from server.resources.metaclasses.all_optional_metaclass import AllOptionalMetaclass
+from server.resources.mixins.timestamp_mixin import TimestampMixin
 from server.resources.people_resources import CreatePeople
 
 
@@ -44,7 +46,7 @@ class UpdateUserOptional(UpdateUser, metaclass=AllOptionalMetaclass):
     pass
 
 
-class User(UpdateUser):
+class User(TimestampMixin, UpdateUser):
     id: int
